// carpool-front/src/app/modules/openapi/model/stats.ts
/**
 * carpool-back API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Family } from './family';


export interface Stats { 
    family?: Family;
    meanTripPerWeek?: number;
    perfectMeanTripPerWeek?: number;
}